+++ examples/save_file/src/main.rs
@@ -37,6 +37,6 @@ fn main() -> eframe::Result<()> {
     eframe::run_native(
         "File dialog example",
         eframe::NativeOptions::default(),
-        Box::new(|ctx| Box::new(MyApp::new(ctx))),
+        Box::new(|ctx| Ok(Box::new(MyApp::new(ctx)))),
     )
 }