+++ src/config/mod.rs
@@ -1,68 +1,38 @@
+mod labels;
+pub use labels::FileDialogLabels;
+
+mod keybindings;
+pub use keybindings::{FileDialogKeyBindings, KeyBinding};
+
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
-/// Function that returns true if the specific item matches the filter.
-pub type Filter<T> = Arc<dyn Fn(&T) -> bool>;
+use crate::data::DirectoryEntry;
 
-/// Sets a specific icon for directory entries.
-#[derive(Clone)]
-pub struct IconFilter {
-    /// The icon that should be used.
-    pub icon: String,
-    /// Sets a filter function that checks whether a given Path matches the criteria for this icon.
-    pub filter: Filter<Path>,
-}
-
-impl std::fmt::Debug for IconFilter {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("IconFilter")
-            .field("icon", &self.icon)
-            .finish()
-    }
-}
-
-/// Stores the display name and the actual path of a quick access link.
+/// Contains data of the FileDialog that should be stored persistently.
 #[derive(Debug, Clone)]
-pub struct QuickAccessPath {
-    pub display_name: String,
-    pub path: PathBuf,
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+pub struct FileDialogStorage {
+    /// The folders the user pinned to the left sidebar.
+    pub pinned_folders: Vec<DirectoryEntry>,
+    /// If hidden files and folders should be listed inside the directory view.
+    pub show_hidden: bool,
 }
 
-/// Stores a custom quick access section of the file dialog.
-#[derive(Debug, Clone)]
-pub struct QuickAccess {
-    pub canonicalize_paths: bool,
-    pub heading: String,
-    pub paths: Vec<QuickAccessPath>,
-}
-
-impl QuickAccess {
-    /// Adds a new path to the quick access.
-    ///
-    /// Since `fs::canonicalize` is used, both absolute paths and relative paths are allowed.
-    /// See `FileDialog::canonicalize_paths` for more information.
-    ///
-    /// See `FileDialogConfig::add_quick_access` for an example.
-    pub fn add_path(&mut self, display_name: &str, path: impl Into<PathBuf>) {
-        let path = path.into();
-
-        let canonicalized_path = match self.canonicalize_paths {
-            true => fs::canonicalize(&path).unwrap_or(path),
-            false => path,
-        };
-
-        self.paths.push(QuickAccessPath {
-            display_name: display_name.to_string(),
-            path: canonicalized_path,
-        });
+impl Default for FileDialogStorage {
+    /// Creates a new object with default values
+    fn default() -> Self {
+        Self {
+            pinned_folders: Vec::new(),
+            show_hidden: false,
+        }
     }
 }
 
 /// Contains configuration values of a file dialog.
 ///
-/// The configuration of a file dialog can be set using
-/// `FileDialog::with_config` or `FileDialog::overwrite_config`.
+/// The configuration of a file dialog can be set using `FileDialog::with_config`.
 ///
 /// If you only need to configure a single file dialog, you don't need to
 /// manually use a `FileDialogConfig` object. `FileDialog` provides setter methods for
@@ -92,13 +62,28 @@ impl QuickAccess {
 #[derive(Debug, Clone)]
 pub struct FileDialogConfig {
     // ------------------------------------------------------------------------
-    // General options:
+    // Core:
+    /// Persistent data of the file dialog.
+    pub storage: FileDialogStorage,
     /// The labels that the dialog uses.
     pub labels: FileDialogLabels,
+    /// Keybindings used by the file dialog.
+    pub keybindings: FileDialogKeyBindings,
+
+    // ------------------------------------------------------------------------
+    // General options:
+    /// If the file dialog should be visible as a modal window.
+    /// This means that the input outside the window is not registered.
+    pub as_modal: bool,
+    /// Color of the overlay that is displayed under the modal to prevent user interaction.
+    pub modal_overlay_color: egui::Color32,
     /// The first directory that will be opened when the dialog opens.
     pub initial_directory: PathBuf,
     /// The default filename when opening the dialog in `DialogMode::SaveFile` mode.
     pub default_file_name: String,
+    /// If the user is allowed to select an already existing file when the dialog is
+    /// in `DialogMode::SaveFile` mode.
+    pub allow_file_overwrite: bool,
     /// Sets the separator of the directories when displaying a path.
     /// Currently only used when the current path is displayed in the top panel.
     pub directory_separator: String,
@@ -107,15 +92,23 @@ pub struct FileDialogConfig {
 
     /// The icon that is used to display error messages.
     pub err_icon: String,
+    /// The icon that is used to display warning messages.
+    pub warn_icon: String,
     /// The default icon used to display files.
     pub default_file_icon: String,
     /// The default icon used to display folders.
     pub default_folder_icon: String,
+    /// The icon used to display pinned paths in the left panel.
+    pub pinned_icon: String,
     /// The icon used to display devices in the left panel.
     pub device_icon: String,
     /// The icon used to display removable devices in the left panel.
     pub removable_device_icon: String,
 
+    /// File filters presented to the user in a dropdown.
+    pub file_filters: Vec<FileFilter>,
+    /// Name of the file filter to be selected by default.
+    pub default_file_filter: Option<String>,
     /// Sets custom icons for different files or folders.
     /// Use `FileDialogConfig::set_file_icon` to add a new icon to this list.
     pub file_icon_filters: Vec<IconFilter>,
@@ -167,14 +160,21 @@ pub struct FileDialogConfig {
     pub show_current_path: bool,
     /// If the button to text edit the current path should be visible.
     pub show_path_edit_button: bool,
-    /// If the reload button in the top panel should be visible.
+    /// If the menu button containing the reload button and other options should be visible.
+    pub show_menu_button: bool,
+    /// If the reload button inside the top panel menu should be visible.
     pub show_reload_button: bool,
+    /// If the show hidden files and folders option inside the top panel menu should be visible.
+    pub show_hidden_option: bool,
     /// If the search input in the top panel should be visible.
     pub show_search: bool,
 
     /// If the sidebar with the shortcut directories such as
     /// “Home”, “Documents” etc. should be visible.
     pub show_left_panel: bool,
+    /// If pinned folders should be listed in the left sidebar.
+    /// Disabling this will also disable the functionality to pin a folder.
+    pub show_pinned_folders: bool,
     /// If the Places section in the left sidebar should be visible.
     pub show_places: bool,
     /// If the Devices section in the left sidebar should be visible.
@@ -187,18 +187,28 @@ impl Default for FileDialogConfig {
     /// Creates a new configuration with default values
     fn default() -> Self {
         Self {
+            storage: FileDialogStorage::default(),
             labels: FileDialogLabels::default(),
+            keybindings: FileDialogKeyBindings::default(),
+
+            as_modal: true,
+            modal_overlay_color: egui::Color32::from_rgba_premultiplied(0, 0, 0, 120),
             initial_directory: std::env::current_dir().unwrap_or_default(),
             default_file_name: String::new(),
+            allow_file_overwrite: true,
             directory_separator: String::from(">"),
             canonicalize_paths: true,
 
             err_icon: String::from("⚠"),
+            warn_icon: String::from("⚠"),
             default_file_icon: String::from("🗋"),
             default_folder_icon: String::from("🗀"),
+            pinned_icon: String::from("📌"),
             device_icon: String::from("🖴"),
             removable_device_icon: String::from("💾"),
 
+            file_filters: Vec::new(),
+            default_file_filter: None,
             file_icon_filters: Vec::new(),
 
             quick_accesses: Vec::new(),
@@ -222,10 +232,13 @@ impl Default for FileDialogConfig {
             show_new_folder_button: true,
             show_current_path: true,
             show_path_edit_button: true,
+            show_menu_button: true,
             show_reload_button: true,
+            show_hidden_option: true,
             show_search: true,
 
             show_left_panel: true,
+            show_pinned_folders: true,
             show_places: true,
             show_devices: true,
             show_removable_devices: true,
@@ -234,6 +247,56 @@ impl Default for FileDialogConfig {
 }
 
 impl FileDialogConfig {
+    /// Sets the storage used by the file dialog.
+    /// Storage includes all data that is persistently stored between multiple
+    /// file dialog instances.
+    pub fn storage(mut self, storage: FileDialogStorage) -> Self {
+        self.storage = storage;
+        self
+    }
+
+    /// Adds a new file filter the user can select from a dropdown widget.
+    ///
+    /// NOTE: The name must be unique. If a filter with the same name already exists,
+    ///       it will be overwritten.
+    ///
+    /// # Arguments
+    ///
+    /// * `name` - Display name of the filter
+    /// * `filter` - Sets a filter function that checks whether a given
+    ///   Path matches the criteria for this filter.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use std::sync::Arc;
+    /// use egui_file_dialog::FileDialogConfig;
+    ///
+    /// let config = FileDialogConfig::default()
+    ///     .add_file_filter(
+    ///         "PNG files",
+    ///         Arc::new(|path| path.extension().unwrap_or_default() == "png"))
+    ///     .add_file_filter(
+    ///         "JPG files",
+    ///         Arc::new(|path| path.extension().unwrap_or_default() == "jpg"));
+    /// ```
+    pub fn add_file_filter(mut self, name: &str, filter: Filter<Path>) -> Self {
+        let id = egui::Id::new(name);
+
+        if let Some(item) = self.file_filters.iter_mut().find(|p| p.id == id) {
+            item.filter = filter.clone();
+            return self;
+        }
+
+        self.file_filters.push(FileFilter {
+            id,
+            name: name.to_string(),
+            filter,
+        });
+
+        self
+    }
+
     /// Sets a new icon for specific files or folders.
     ///
     /// # Arguments
@@ -293,122 +356,83 @@ impl FileDialogConfig {
     }
 }
 
-/// Contains the text labels that the file dialog uses.
-///
-/// This is used to enable multiple language support.
-///
-/// # Example
-///
-/// The following example shows how the default title of the dialog can be displayed
-/// in German instead of English.
-///
-/// ```
-/// use egui_file_dialog::{FileDialog, FileDialogLabels};
-///
-/// let labels_german = FileDialogLabels {
-///     title_select_directory: "📁 Ordner Öffnen".to_string(),
-///     title_select_file: "📂 Datei Öffnen".to_string(),
-///     title_save_file: "📥 Datei Speichern".to_string(),
-///     ..Default::default()
-/// };
-///
-/// let file_dialog = FileDialog::new().labels(labels_german);
-/// ```
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct FileDialogLabels {
-    // ------------------------------------------------------------------------
-    // General:
-    /// The default window title used when the dialog is in `DialogMode::SelectDirectory` mode.
-    pub title_select_directory: String,
-    /// The default window title used when the dialog is in `DialogMode::SelectFile` mode.
-    pub title_select_file: String,
-    /// The default window title used when the dialog is in `DialogMode::SaveFile` mode.
-    pub title_save_file: String,
+/// Function that returns true if the specific item matches the filter.
+pub type Filter<T> = Arc<dyn Fn(&T) -> bool + Send + Sync>;
 
-    // ------------------------------------------------------------------------
-    // Left panel:
-    /// Heading of the "Places" section in the left panel
-    pub heading_places: String,
-    /// Heading of the "Devices" section in the left panel
-    pub heading_devices: String,
-    /// Heading of the "Removable Devices" section in the left panel
-    pub heading_removable_devices: String,
-
-    /// Name of the home directory
-    pub home_dir: String,
-    /// Name of the desktop directory
-    pub desktop_dir: String,
-    /// Name of the documents directory
-    pub documents_dir: String,
-    /// Name of the downloads directory
-    pub downloads_dir: String,
-    /// Name of the audio directory
-    pub audio_dir: String,
-    /// Name of the pictures directory
-    pub pictures_dir: String,
-    /// Name of the videos directory
-    pub videos_dir: String,
-
-    // ------------------------------------------------------------------------
-    // Bottom panel:
-    /// Text that appears in front of the selected folder preview in the bottom panel.
-    pub selected_directory: String,
-    /// Text that appears in front of the selected file preview in the bottom panel.
-    pub selected_file: String,
-    /// Text that appears in front of the file name input in the bottom panel.
-    pub file_name: String,
-
-    /// Button text to open the selected item.
-    pub open_button: String,
-    /// Button text to save the file.
-    pub save_button: String,
-    /// Button text to cancel the dialog.
-    pub cancel_button: String,
-
-    // ------------------------------------------------------------------------
-    // Error message:
-    /// Error if no folder name was specified.
-    pub err_empty_folder_name: String,
-    /// Error if no file name was specified.
-    pub err_empty_file_name: String,
-    /// Error if the directory already exists.
-    pub err_directory_exists: String,
-    /// Error if the file already exists.
-    pub err_file_exists: String,
+/// Defines a specific file filter that the user can select from a dropdown.
+#[derive(Clone)]
+pub struct FileFilter {
+    /// The ID of the file filter, used internally for identification.
+    pub id: egui::Id,
+    /// The display name of the file filter
+    pub name: String,
+    /// Sets a filter function that checks whether a given Path matches the criteria for this file.
+    pub filter: Filter<Path>,
 }
 
-impl Default for FileDialogLabels {
-    /// Creates a new object with the default english labels.
-    fn default() -> Self {
-        Self {
-            title_select_directory: "📁 Select Folder".to_string(),
-            title_select_file: "📂 Open File".to_string(),
-            title_save_file: "📥 Save File".to_string(),
+impl std::fmt::Debug for FileFilter {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("FileFilter")
+            .field("name", &self.name)
+            .finish()
+    }
+}
 
-            heading_places: "Places".to_string(),
-            heading_devices: "Devices".to_string(),
-            heading_removable_devices: "Removable Devices".to_string(),
+/// Sets a specific icon for directory entries.
+#[derive(Clone)]
+pub struct IconFilter {
+    /// The icon that should be used.
+    pub icon: String,
+    /// Sets a filter function that checks whether a given Path matches the criteria for this icon.
+    pub filter: Filter<Path>,
+}
 
-            home_dir: "🏠  Home".to_string(),
-            desktop_dir: "🖵  Desktop".to_string(),
-            documents_dir: "🗐  Documents".to_string(),
-            downloads_dir: "📥  Downloads".to_string(),
-            audio_dir: "🎵  Audio".to_string(),
-            pictures_dir: "🖼  Pictures".to_string(),
-            videos_dir: "🎞  Videos".to_string(),
+impl std::fmt::Debug for IconFilter {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("IconFilter")
+            .field("icon", &self.icon)
+            .finish()
+    }
+}
 
-            selected_directory: "Selected directory:".to_string(),
-            selected_file: "Selected file:".to_string(),
-            file_name: "File name:".to_string(),
+/// Stores the display name and the actual path of a quick access link.
+#[derive(Debug, Clone)]
+pub struct QuickAccessPath {
+    /// Name of the path that is shown inside the left panel.
+    pub display_name: String,
+    /// Absolute or relative path to the folder.
+    pub path: PathBuf,
+}
 
-            open_button: "🗀  Open".to_string(),
-            save_button: "📥  Save".to_string(),
-            cancel_button: "🚫 Cancel".to_string(),
+/// Stores a custom quick access section of the file dialog.
+#[derive(Debug, Clone)]
+pub struct QuickAccess {
+    /// If the path's inside the quick access section should be canonicalized.
+    canonicalize_paths: bool,
+    /// Name of the quick access section displayed inside the left panel.
+    pub heading: String,
+    /// Path's contained inside the quick access section.
+    pub paths: Vec<QuickAccessPath>,
+}
 
-            err_empty_folder_name: "Name of the folder cannot be empty".to_string(),
-            err_empty_file_name: "The file name cannot be empty".to_string(),
-            err_directory_exists: "A directory with the name already exists".to_string(),
-            err_file_exists: "A file with the name already exists".to_string(),
-        }
+impl QuickAccess {
+    /// Adds a new path to the quick access.
+    ///
+    /// Since `fs::canonicalize` is used, both absolute paths and relative paths are allowed.
+    /// See `FileDialog::canonicalize_paths` for more information.
+    ///
+    /// See `FileDialogConfig::add_quick_access` for an example.
+    pub fn add_path(&mut self, display_name: &str, path: impl Into<PathBuf>) {
+        let path = path.into();
+
+        let canonicalized_path = match self.canonicalize_paths {
+            true => fs::canonicalize(&path).unwrap_or(path),
+            false => path,
+        };
+
+        self.paths.push(QuickAccessPath {
+            display_name: display_name.to_string(),
+            path: canonicalized_path,
+        });
     }
 }