+++ src/data/directory_content.rs
@@ -1,17 +1,20 @@
 use std::path::{Path, PathBuf};
 use std::{fs, io};
 
-use crate::FileDialogConfig;
+use crate::config::{FileDialogConfig, FileFilter};
 
 /// Contains the metadata of a directory item.
 /// This struct is mainly there so that the metadata can be loaded once and not that
 /// a request has to be sent to the OS every frame using, for example, `path.is_file()`.
-#[derive(Default, Clone, PartialEq, Eq)]
+#[derive(Debug, Default, Clone)]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub struct DirectoryEntry {
     path: PathBuf,
     is_directory: bool,
     is_system_file: bool,
     icon: String,
+    /// If the item is marked as selected as part of a multi selection.
+    pub selected: bool,
 }
 
 impl DirectoryEntry {
@@ -22,7 +25,13 @@ impl DirectoryEntry {
             is_directory: path.is_dir(),
             is_system_file: !path.is_dir() && !path.is_file(),
             icon: gen_path_icon(config, path),
+            selected: false,
         }
+    }
+
+    /// Checks if the path of the current directory entry matches the other directory entry.
+    pub fn path_eq(&self, other: &DirectoryEntry) -> bool {
+        other.as_path() == self.as_path()
     }
 
     /// Returns true if the item is a directory.
@@ -50,6 +59,11 @@ impl DirectoryEntry {
     }
 
     /// Returns the path of the directory item.
+    pub fn as_path(&self) -> &Path {
+        &self.path
+    }
+
+    /// Clones the path of the directory item.
     pub fn to_path_buf(&self) -> PathBuf {
         self.path.clone()
     }
@@ -88,6 +102,11 @@ impl DirectoryEntry {
                 ""
             })
     }
+
+    /// Returns whether the path this DirectoryEntry points to is considered hidden.
+    pub fn is_hidden(&self) -> bool {
+        is_path_hidden(self)
+    }
 }
 
 /// Contains the content of a directory.
@@ -114,15 +133,76 @@ impl DirectoryContent {
         })
     }
 
-    /// Very simple wrapper methods of the contents .iter() method.
-    /// No trait is implemented since this is currently only used internal.
-    pub fn iter(&self) -> std::slice::Iter<'_, DirectoryEntry> {
-        self.content.iter()
+    /// Checks if the given directory entry is visible with the applied filters.
+    fn is_entry_visible(
+        dir_entry: &DirectoryEntry,
+        show_hidden: bool,
+        search_value: &str,
+        file_filter: Option<&FileFilter>,
+    ) -> bool {
+        if !search_value.is_empty()
+            && !dir_entry
+                .file_name()
+                .to_lowercase()
+                .contains(&search_value.to_lowercase())
+        {
+            return false;
+        }
+
+        if !show_hidden && dir_entry.is_hidden() {
+            return false;
+        }
+
+        if let Some(file_filter) = file_filter {
+            if dir_entry.is_file() && !(file_filter.filter)(dir_entry.as_path()) {
+                return false;
+            }
+        }
+
+        true
+    }
+
+    pub fn filtered_iter<'s>(
+        &'s self,
+        show_hidden: bool,
+        search_value: &'s str,
+        file_filter: Option<&'s FileFilter>,
+    ) -> impl Iterator<Item = &DirectoryEntry> + 's {
+        self.content
+            .iter()
+            .filter(move |p| Self::is_entry_visible(p, show_hidden, search_value, file_filter))
+    }
+
+    pub fn filtered_iter_mut<'s>(
+        &'s mut self,
+        show_hidden: bool,
+        search_value: &'s str,
+        file_filter: Option<&'s FileFilter>,
+    ) -> impl Iterator<Item = &mut DirectoryEntry> + 's {
+        self.content
+            .iter_mut()
+            .filter(move |p| Self::is_entry_visible(p, show_hidden, search_value, file_filter))
+    }
+
+    pub fn reset_multi_selection(&mut self) {
+        for item in self.content.iter_mut() {
+            item.selected = false;
+        }
+    }
+
+    /// Returns the number of elements inside the directory.
+    pub fn len(&self) -> usize {
+        self.content.len()
     }
 
     /// Pushes a new item to the content.
     pub fn push(&mut self, item: DirectoryEntry) {
         self.content.push(item);
+    }
+
+    /// Clears the items inside the directory.
+    pub fn clear(&mut self) {
+        self.content.clear();
     }
 }
 
@@ -162,9 +242,26 @@ fn load_directory(
         },
     });
 
-    // TODO: Implement "Show hidden files and folders" option
-
     Ok(result)
+}
+
+#[cfg(windows)]
+fn is_path_hidden(item: &DirectoryEntry) -> bool {
+    use std::os::windows::fs::MetadataExt;
+
+    match fs::metadata(item.as_path()) {
+        Ok(metadata) => metadata.file_attributes() & 0x2 > 0,
+        Err(_) => false,
+    }
+}
+
+#[cfg(not(windows))]
+fn is_path_hidden(item: &DirectoryEntry) -> bool {
+    if item.file_name().bytes().next() == Some(b'.') {
+        return true;
+    }
+
+    false
 }
 
 /// Generates the icon for the specific path.